README.
I recommend keeping a README of each of your scripts in this file.  For a large enough project, you'd have one README for the entire project, but since this is a culmination of scripts, simple sections with instructions on how to build, run, and use each script would be helpful for whoever comes across this.